+++ apps/desktop/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [0.7.3](https://github.com/JoaoSipauba/leviticus/compare/v0.7.2...v0.7.3) (2026-05-15)
+
+### Correções
+
+* **add-song:** parar prévia ao trocar de aba pesquisa↔URL ([#20](https://github.com/JoaoSipauba/leviticus/issues/20)) ([72f263e](https://github.com/JoaoSipauba/leviticus/commit/72f263ef89b163dbab178ef0b547d0d25efdda44))
+
+## [0.7.2](https://github.com/JoaoSipauba/leviticus/compare/v0.7.1...v0.7.2) (2026-05-14)
+
+### Correções
+
+* **youtube:** search via HTML scrape + binários robustos ([#19](https://github.com/JoaoSipauba/leviticus/issues/19)) ([0dd3973](https://github.com/JoaoSipauba/leviticus/commit/0dd3973afb0988b079c7c9fdad344e285faff2de))
+
+## [0.7.1](https://github.com/JoaoSipauba/leviticus/compare/v0.7.0...v0.7.1) (2026-05-14)
+
+### Correções
+
+* **youtube:** integração 10-20× mais rápida (oEmbed + Innertube) ([#18](https://github.com/JoaoSipauba/leviticus/issues/18)) ([506f8a5](https://github.com/JoaoSipauba/leviticus/commit/506f8a5ff56bbb51281f8ff65fd6d74bf1f2b22c))
+
 ## [0.7.0](https://github.com/JoaoSipauba/leviticus/compare/v0.6.0...v0.7.0) (2026-05-14)
 
 ### Funcionalidades